// src/app/auth/models/transaction.module.ts
export class Transaction {
    transactionDate: string;
    transactionRemarks: string;
    transactionType: string;
    balanceOnDate: string;
    transactionAmount: string;
    fromAccountNumber: string;
    toAccountNumber: string;
    transactionTimeStamp: string;
    transactionId: string;
    transactionStatus: string;
}